Add files via upload
This has a the battle class. it won't be called from the main screen but the basic frame work for the game. a lot of work to be done but overall just repeating code with a different name. I need sprite and game character and enemy sheets. once those are loaded in I can load in a bunch of that into the needed background information.